Move Validateconfig to configtest
